fix: implement true lazy-loading for code of Documentation component
