+++ src/components/SideTools/Documentation.tsx
@@ -2,47 +2,29 @@ import classes from './Documentation.module.css';
 import useDocumentation from './useDocumentation';
 
 const Documentation = () => {
-  const {
-    documentation,
-    handleDocsBtnClick,
-    docsBtnText,
-    isFetching,
-    isContentVisible,
-  } = useDocumentation();
+  const { documentation, handleDocsBtnClick, docsBtnText, isContentVisible } =
+    useDocumentation();
 
-  const renderContent = () => {
-    return isFetching ? (
-      <section className={classes.wrapper}>
-        <div className={classes.documentationTools}>
-          <button onClick={handleDocsBtnClick} disabled>
-            {docsBtnText}
-          </button>
-        </div>
-      </section>
-    ) : (
-      <section
-        className={`${classes.wrapper} ${
+  return (
+    <section
+      className={`${classes.wrapper} ${isContentVisible ? classes.active : ''}`}
+    >
+      <div className={classes.documentationTools}>
+        <button onClick={handleDocsBtnClick}>{docsBtnText}</button>
+      </div>
+      <div
+        className={`${classes.documentationContentWrap} ${
           isContentVisible ? classes.active : ''
         }`}
       >
-        <div className={classes.documentationTools}>
-          <button onClick={handleDocsBtnClick}>{docsBtnText}</button>
-        </div>
-        <div
-          className={`${classes.documentationContentWrap} ${
-            isContentVisible ? classes.active : ''
-          }`}
-        >
-          <textarea
-            value={documentation}
-            className={classes.documentationContent}
-            readOnly
-          />
-        </div>
-      </section>
-    );
-  };
-  return renderContent();
+        <textarea
+          value={documentation}
+          className={classes.documentationContent}
+          readOnly
+        />
+      </div>
+    </section>
+  );
 };
 
 export default Documentation;
