+++ src/pages/MainPage/MainPage.tsx
@@ -4,19 +4,43 @@ import { useAppDispatch, useAppSelector } from '../../hook/useRedux';
 import {
   fetchJSON,
   setApiLink,
+  setDocumentation,
   setError,
+  setIsDocsLoaded,
   setQuery,
 } from '../../redux/slices/GraphQLSlice';
 import Toast from '../../components/Toast/Toast';
 import errorIcon from '../../assets/error-icon.svg';
 import EditorViewerSwitch from '../../components/EditorViewerSwitch/EditorViewerSwitch';
 import EditorTools from '../../components/EditorTools/EditorTools';
-import Documentation from '../../components/SideTools/Documentation';
+import { Suspense, lazy, useEffect } from 'react';
+import { fetchSchema } from '../../utils/fetchSchema';
+
+const Documentation = lazy(
+  () => import('../../components/SideTools/Documentation')
+);
 
 function MainPage() {
-  const { apiLink, query, jsonViewer, error, variables, headers } =
-    useAppSelector((store) => store.graphQL);
+  const {
+    apiLink,
+    query,
+    jsonViewer,
+    error,
+    variables,
+    headers,
+    isDocsLoaded,
+  } = useAppSelector((store) => store.graphQL);
   const dispatch = useAppDispatch();
+
+  useEffect(() => {
+    (async () => {
+      dispatch(setIsDocsLoaded(false));
+      const data = await fetchSchema({ url: apiLink });
+      const jsonData = JSON.stringify(data, null, 4);
+      dispatch(setDocumentation(jsonData));
+      dispatch(setIsDocsLoaded(true));
+    })();
+  }, [apiLink, dispatch]);
 
   const handleErrToastClose = () => {
     dispatch(setError(null));
@@ -42,7 +66,7 @@ function MainPage() {
         </Toast>
       )}
       <div className={classes.row}>
-        <Documentation />
+        <Suspense>{isDocsLoaded && <Documentation />}</Suspense>
         <section className={classes.content}>
           <ApiLinkForm
             value={apiLink}
